Fix menu tests for main
