+++ menus/main_tests.py
@@ -1,35 +1,39 @@
-from constants import menu_options
-from menus.medium_setup import display, get_option_callbacks_map
-from mock import Mock, patch
+from constants.display import medium_setup
+from menus.main import display
+from mock import patch
 
 
-@patch('menus.medium_setup.get_option_callbacks_map')
-@patch('utils.menus.generate_menu')
-def test_display(generate_menu, get_options):
+@patch('click.echo')
+@patch('menus.medium_setup.display')
+@patch('utils.settings.config_exists', return_value=True)
+def test_display_with_existing_config(config_exists,
+                                      display_medium_menu,
+                                      click_echo):
     """
     display
     """
-    options_map = 'Who cares'
-    previous_menu_callback = 'No one.'
-    get_options.return_value = options_map
+    display()
 
-    display(previous_menu_callback)
+    # It should not display the medium setup intro.
+    click_echo.assert_not_called()
 
-    # It should get callbacks from the get_option_callbacks_map function.
-    get_options.assert_called_once_with(previous_menu_callback)
-
-    # It should generate a menu.
-    generate_menu.assert_called_once_with(menu_options.SETUP_MEDIUM,
-                                          options_map,
-                                          display)
+    # It should not show the medium setup menu.
+    display_medium_menu.assert_not_called()
 
 
-def test_get_option_callbacks_map():
+@patch('click.echo')
+@patch('menus.medium_setup.display')
+@patch('utils.settings.config_exists', return_value=False)
+def test_display_without_existing_config(config_exists,
+                                         display_medium_menu,
+                                         click_echo):
     """
-    get_option_callbacks_map
+    display
     """
-    result = get_option_callbacks_map(Mock())
+    display()
 
-    # It should return an options map.
-    for key, value in result.items():
-        assert callable(value)
+    # It should display the medium setup intro.
+    click_echo.assert_called_once_with(medium_setup.INTRO)
+
+    # It should not show the medium setup menu.
+    display_medium_menu.assert_called_once_with(display)
